Chore(dom): DOM - helpers refactor (#165)
